[DESIGN] : 🎨 채팅 레이아웃 수정 및 Adapter ViewType 적용
