+++ app/src/main/java/com/sandamso/sansaninfo/ChattingPage/ChatRoomActivity.kt
@@ -44,7 +44,8 @@ class ChatRoomActivity:AppCompatActivity() {
         roomTitle = intent.getStringExtra("dataFromdetailPageTitle") ?: ""
 
         msgRef = FBRoom.roomRef.child(roomId).child("messages")
-        adapter = MsgListAdapter(msgList)
+        adapter = MsgListAdapter(msgList, currentUser)
+        Log.d("currentUser", "$currentUser")
 
         with(binding){
             recyclerMessages.adapter = adapter
