+++ app/src/main/res/layout/list_talk_item_mine.xml
@@ -8,29 +8,18 @@
     android:paddingVertical="15dp">
 
     <TextView
-        android:id="@+id/txt_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginEnd="10dp"
-        android:textStyle="bold"
-        android:text="날다람쥐"
-        app:layout_constraintBottom_toTopOf="@id/txt_message"/>
-
-    <TextView
         android:id="@+id/txt_message"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:autoLink="web"
         android:background="@drawable/background_talk_mine"
         android:linksClickable="true"
+        android:layout_marginEnd="10dp"
         android:padding="10dp"
         android:text="안녕하세요."
         android:textColor="@color/black"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/txt_name"
-        android:layout_marginTop="5dp" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/txt_date"
