+++ app/src/main/res/layout/list_talk_item_others.xml
@@ -9,6 +9,16 @@
     android:layout_marginBottom="20dp">
 
     <TextView
+        android:id="@+id/txt_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="15dp"
+        android:text="날다람쥐"
+        android:textStyle="bold"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
         android:id="@+id/txt_message"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -16,10 +26,12 @@
         android:background="@drawable/background_talk_others"
         android:linksClickable="true"
         android:padding="10dp"
+        android:layout_marginStart="10dp"
         android:text="안녕하세요."
         android:textColor="@color/white"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/txt_name"
+        android:layout_marginTop="5dp"/>
 
     <TextView
         android:id="@+id/txt_date"
